requirements.txt
click==8.1.3
psycopg2-binary
ujson==5.4.0
requests
wand
Flask==2.1.2
